Rename 'WordClock++' to 'RELOK'
