--- android/res/drawable/splashscreen_port.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="#000000"/>
-        </shape>
-    </item>
-    <item>
-        <bitmap android:src="@drawable/logo_port" android:gravity="center"/>
-    </item>
-</layer-list>
