+++ android/res/values/splashscreentheme.xml
@@ -2,7 +2,5 @@
 <resources>
     <style name="splashScreenTheme">
         <item name="android:windowBackground">@drawable/splashscreen</item>
-        <item name="android:windowSplashScreenAnimatedIcon">@android:color/transparent</item>
-        <item name="android:windowSplashScreenAnimationDuration">0</item>
     </style>
 </resources>
