+++ fastlane/Fastfile
@@ -54,7 +54,7 @@ platform :ios do
     # Fetch read your app identifier defined in your Appfile
     app_identifier = CredentialsManager::AppfileConfig.try_fetch_value(:app_identifier)
     api_key = lane_context[SharedValues::APP_STORE_CONNECT_API_KEY]
-    current_version = get_version_number(target: "WordClock")
+    current_version = get_version_number(target: "RELOK")
     latest_build_number = latest_testflight_build_number(
       api_key: api_key,
       version: current_version,
@@ -89,13 +89,13 @@ platform :ios do
   lane :build_release do |options|
     app_identifier = CredentialsManager::AppfileConfig.try_fetch_value(:app_identifier)
     profile_name = "App Provisioning Profile"
-    output_name = "wordclock"
+    output_name = "RELOK"
     export_method = "app-store"
 
     # Turn off automatic signing during build so the correct code signing identity is guaranteed to be used
     update_code_signing_settings(
       use_automatic_signing: false,
-      targets: ["WordClock"],
+      targets: ["RELOK"],
       code_sign_identity: "Apple Distribution",
       bundle_identifier: app_identifier,
       profile_name: profile_name,
@@ -103,13 +103,13 @@ platform :ios do
     )
 
     update_project_team(
-      path: "WordClock.xcodeproj",
+      path: "RELOK.xcodeproj",
       teamid: ENV["TEAM_ID"]
     )
 
     # Build the app
     build_app(
-      scheme: "WordClock",
+      scheme: "RELOK",
       output_name: output_name,
       configuration: "Release",
       export_options: {
