+++ src/main.cpp
@@ -28,7 +28,7 @@ int main(int argc, char *argv[])
     QGuiApplication app(argc, argv);
     app.setOrganizationName("Kokleeko S.L.");
     app.setOrganizationDomain("kokleeko.io");
-    app.setApplicationName("WordClock++");
+    app.setApplicationName("RELOK");
     app.setApplicationVersion(VERSION);
     QQmlApplicationEngine engine;
