+++ wasm/app/index.html
@@ -9,7 +9,7 @@
         and Qt device independent pixels. -->
     <meta name="viewport" content="width=device-width, height=device-height, user-scalable=0"/>
 
-    <title>WordClock++ - Web Application</title>
+    <title>RELOK - Web Application</title>
     <link rel="icon" href="/logo.svg" type="image/svg+xml">
     <meta name="description" content="web application built with Qt that displays time as a word clock">
     <style>
@@ -61,7 +61,7 @@
                   canvas.style.display = 'none';
               },
           });
-          qtLoader.loadEmscriptenModule("WordClock");
+          qtLoader.loadEmscriptenModule("RELOK");
       }
     </script>
     <script type="text/javascript" src="qtloader.js"></script>
