+++ webos/appinfo.json
@@ -4,7 +4,7 @@
   "vendor": "Kokleeko S.L",
   "type": "qml",
   "main": "main.qml",
-  "title": "WordClock",
+  "title": "RELOK",
   "icon": "icon.png",
   "requiredPermissions": [
     "time.query",
